+++ tsconfig.json
@@ -8,18 +8,18 @@
       "noImplicitAny": false,
       "moduleResolution": "node",
       "baseUrl": ".",                       
-      "paths": {                                 
-           "mattermost/*": ["src/mattermost/*"],
-           "zendesk/*": ["./src/zendesk/*"],
-           "src/*": ["./src/*"]
-       },
-       "typeRoots": [
-         "./src"
-       ], 
-       "types": [
-         "jest",
-         "node"
-       ], 
+      "paths": {
+         "mattermost/*": ["./src/mattermost/*"],
+         "zendesk/*": ["./src/zendesk/*"],
+         "src/*": ["./src/*"]
+      },
+      "typeRoots": [
+        "./src"
+      ], 
+      "types": [
+        "jest",
+        "node"
+      ], 
       "esModuleInterop": true,
       "skipLibCheck": true
     },
